refactor: isolate runtime retention inventory
